Clean up types and constants
