+++ lib/constants-server.ts
@@ -1,4 +1,4 @@
-import { PopulationState, ScenarioGroupState } from './types';
+// Some constants for the server only.
 export const USER_KEY_COOKIE = 'userKey';
 
 export const RES_MSG = {
@@ -23,36 +23,4 @@ export const RES_MSG = {
   NO_POPULATION_WITH_ID: 'There is no population with the provided populationId.',
   NO_SCENARIO_GROUP_WITH_ID: 'There is no scenario group with the given scenarioGroupId.',
   NO_SCENARIO_WITH_ID: 'There is no scenario with the provided scenarioId.',
-};
-
-export const NULL_POPULATION_STATE: PopulationState = {
-  loading: false,
-  population: {
-    populationId: 'ckrjtocix00950745zrvans2v',
-    name: '',
-    ownerId: '',
-    elements: [],
-  },
-};
-
-export const NULL_SUCCESSES_STATE: ScenarioGroupState = {
-  loading: false,
-  scenarioGroup: {
-    scenarioGroupId: '',
-    name: '',
-    type: 'SUCCESSES',
-    populationId: '',
-    scenarios: [],
-  },
-};
-
-export const NULL_FAILURES_STATE: ScenarioGroupState = {
-  loading: false,
-  scenarioGroup: {
-    scenarioGroupId: '',
-    name: '',
-    type: 'FAILURES',
-    populationId: '',
-    scenarios: [],
-  },
 };
